+++ Lemonade Stand/Assets/Scripts/PlayerStats.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+
+public class PlayerStats
+{
+    public int Attraction { get; set; }
+
+    public PlayerStats() {
+        Attraction = 10;
+    }
+}