+++ Lemonade Stand/Assets/Scripts/PurchaseRequest.cs
@@ -1,0 +1,11 @@
+public class PurchaseRequest {
+    public ItemType ItemType { get; }
+    public decimal Price { get; }
+    public int Amount { get; }
+
+    public PurchaseRequest(ItemType itemType, decimal price, int amount) {
+        ItemType = itemType;
+        Price = price;
+        Amount = amount;
+    }
+}