+++ Lemonade Stand/Assets/Scripts/SupplyShop.cs
@@ -1,11 +1,17 @@
+using System;
 using UnityEngine;
 
-public class SupplyShop
-{
+public class SupplyShop : MonoBehaviour {
     const decimal lemon_price = 0.5m;
 
-    public decimal GetLemonPrice(int amount)
-    {
-        return lemon_price * amount;
+    public static event Action<PurchaseRequest> OnPurchaseRequested;
+
+    private void Start() {
+        UIButtonListener.OnSupplyShopButtonClicked += BuyLemons;
+    }
+
+    void BuyLemons() {
+        PurchaseRequest purchaseRequest = new PurchaseRequest(ItemType.Lemon, lemon_price, 1);
+        OnPurchaseRequested?.Invoke(purchaseRequest);
     }
 }