+++ Lemonade Stand/Assets/Scripts/World.cs
@@ -1,14 +1,17 @@
 using System;
-using UnityEngine.LightTransport;
+using UnityEngine;
 
-public class World {
+public class World : MonoBehaviour {
     public static event Action<int> OnDayCountChanged;
     public static event Action<float> OnDayTimerTicked;
+    public static event Action<float> OnEndEarlyTimerTicked;
+    public static event Action OnDayStart;
+    public static event Action OnDayEnd;
 
     private int _day_count;
 
     public Timer DayTimer { get; private set; }
-
+    public Timer EndEarlyTimer { get; private set; }
     public int DayCount {
         get => _day_count;
         private set {
@@ -19,36 +22,54 @@ public class World {
 
     public const float day_duration = 180f;// how many seconds is one day, set to 180s
     public const float end_early_timer_duration = 5f; // how long end early timer goes, set to 5s
-    public bool day_running; // determines if day is running
-    public Timer end_early_timer; // timer for ending day early if no customers
 
-    public World()
-    {
-        DayCount = 1; // start day count at 1
-        day_running = false; // start with day not running
-        DayTimer = new Timer(day_duration); // set day timer to day duration
-        end_early_timer = new Timer(end_early_timer_duration); // set timer to end early duration
+    private void Start() {
+        DayCount = 1; // start at first day
+        CreateTimers();
+
+        UIButtonListener.OnStartButtonClicked += StartDay;
+        CustomerManager.OnNoCustomers += EndEarly;
+    }
+
+    private void Update() {
+        DayTimer.Tick();
+        EndEarlyTimer.Tick();
+    }
+
+    void StopUpdating() {
+        enabled = false;
+    }
+
+    void StartUpdating() {
+        enabled = true;
+    }
+
+    public void CreateTimers() {
+        DayTimer = new Timer(day_duration);
+        EndEarlyTimer = new Timer(end_early_timer_duration);
 
         DayTimer.OnTimerTicked += OnDayTimerTicked;
+        EndEarlyTimer.OnTimerTicked += OnEndEarlyTimerTicked;
+        DayTimer.OnTimerElapsed += EndDay;
+        EndEarlyTimer.OnTimerElapsed += EndDay;
     }
 
     public void StartDay()
     {
-        DayTimer = new Timer(day_duration);
-        DayTimer.OnTimerTicked += OnDayTimerTicked;
+        CreateTimers();
         DayTimer.Start();
-        end_early_timer = new Timer(end_early_timer_duration);
-
-        day_running = true;
+        StartUpdating();
+        OnDayStart?.Invoke();
     }
 
-    public void Update()
-    {
-        DayTimer.Tick();
-        end_early_timer.Tick();
+    void EndDay() {
+        RaiseDayCounter();
+        StopUpdating();
+        OnDayEnd?.Invoke();
+    }
 
-        // end day if day timer or end early timer is up
-        if (DayTimer.elapsed || end_early_timer.elapsed) day_running = false;
+    public void EndEarly() {
+        EndEarlyTimer.Start();
     }
 
     public void RaiseDayCounter() => DayCount++;